Refactor: seperate `fc2util.ParseNumber` (#149)
